Fix: Interrupted UPDATE/CREATE with DELETE would cause error and orphan state
Observed the following sequence:

1) Create long-running stack create
2) During CREATE_IN_PROGRESS, delete k8s Stack
3) Delete timestamps got added, no deleteStack call made to AWS
4) Stack completed
5) Operator stopped following, but left resource as CREATE_IN_PROGRESS

The problem was in updates being performed to an out-of-date object handle (in the follower). Now, it fetches a fresh object (cached) for the latest version we have. This will give it more opportunity to work itself out.

See also: https://github.com/cuppett/cloudformation-operator/pull/1#discussion_r601849772